Deplace logic that makes the mapping to TrainingService
