+++ src/Showcase/src/Smart.FA.Catalog.Showcase.Web/Services/Training/TrainingService.cs
@@ -1,7 +1,11 @@
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
 using Smart.FA.Catalog.Shared.Collections;
 using Smart.FA.Catalog.Showcase.Infrastructure.Data;
 using Smart.FA.Catalog.Showcase.Infrastructure.Data.Extensions;
 using Smart.FA.Catalog.Showcase.Web.Mappers;
+using Smart.FA.Catalog.Showcase.Web.Options;
+using Smart.FA.Catalog.Showcase.Web.Pages.Training.TrainingDetails;
 using Smart.FA.Catalog.Showcase.Web.Pages.Training.TrainingList;
 
 namespace Smart.FA.Catalog.Showcase.Web.Services.Training;
@@ -10,9 +14,12 @@ public class TrainingService : ITrainingService
 {
     private readonly CatalogShowcaseContext _catalogShowcaseContext;
 
-    public TrainingService(CatalogShowcaseContext catalogShowcaseContext)
+    private readonly MinIOOptions _minIoSettings;
+
+    public TrainingService(CatalogShowcaseContext catalogShowcaseContext, IOptions<MinIOOptions> minIOSettings)
     {
         _catalogShowcaseContext = catalogShowcaseContext;
+        _minIoSettings = minIOSettings.Value ?? throw new InvalidOperationException($"{nameof(minIOSettings)} options not defined");
     }
 
     public async Task<PagedList<TrainingListViewModel>> GetPaginatedTrainingViewModelsAsync(int pageNumber, int pageSize, bool random = false)
@@ -43,4 +50,43 @@ public class TrainingService : ITrainingService
 
         return new PagedList<TrainingListViewModel>(trainings.ToTrainingListViewModels(), new PageItem(currentPage, pageSize), trainings.TotalCount);
     }
+
+    public async Task<TrainingDetailsViewModel?> GetTrainingDetailsViewModelsByIdAsync(int trainingId)
+    {
+        var trainingDetails = await _catalogShowcaseContext.TrainingDetails.Where(training => training.Id == trainingId).ToListAsync();
+
+        if (!trainingDetails.Any())
+        {
+            return null;
+        }
+
+        return MapTrainingDetails(trainingDetails);
+    }
+
+    private TrainingDetailsViewModel MapTrainingDetails(List<Domain.Models.TrainingDetails> trainingDetails)
+    {
+        if (trainingDetails is null)
+        {
+            return null;
+        }
+
+        var firstTrainerDetail = trainingDetails.FirstOrDefault();
+        return new TrainingDetailsViewModel
+        {
+            Id = firstTrainerDetail.Id,
+            TrainingTitle = firstTrainerDetail.TrainingTitle,
+            Goal = firstTrainerDetail.Goal ?? string.Empty,
+            Methodology = firstTrainerDetail.Methodology ?? string.Empty,
+            PracticalModalities = firstTrainerDetail.PracticalModalities ?? string.Empty,
+            TrainerFirstName = firstTrainerDetail.TrainerFirstName,
+            TrainerLastName = firstTrainerDetail.TrainerLastName,
+            TrainerId = firstTrainerDetail.TrainerId,
+            TrainerTitle = firstTrainerDetail.TrainerTitle,
+            Status = TrainingStatusType.FromValue(firstTrainerDetail.StatusId),
+            Topics = trainingDetails.Select(x => Topic.FromValue(x.TrainingTopicId)).ToList(),
+            Languages = trainingDetails.Select(x => x.Language).Distinct().ToList(),
+            TrainerProfileImageUrl = _minIoSettings.GenerateMinIoTrainerProfileUrl(firstTrainerDetail.ProfileImagePath)
+        };
+    }
+
 }
